Copied de.js contents back from de-fix.js
